fix: 빌드 시 타입 오류 해결
