--- src/types/InvitationData.type.ts
@@ -1,3 +1,0 @@
-import { Database } from './database.types';
-
-export type Invitation = Database['public']['Tables']['invitation']['Row'];
