--- src/types/reviewData.types.ts
@@ -1,3 +1,0 @@
-import { Database } from './database.types';
-
-export type Review = Database['public']['Tables']['reviews']['Row'];
