+++ src/utils/supabase/client.ts
@@ -1,8 +1,7 @@
 import { createBrowserClient } from '@supabase/ssr';
-import { Database } from '@/types/database.types';
 
 export const createClient = () =>
-  createBrowserClient<Database>(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!);
+  createBrowserClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!);
 
 const browserClient = createClient();
